fix: remove size from configure
Setting an explicit size when calling configure() on a GPUCanvasContext has been deprecated, and will soon be removed. Please set the canvas width and height attributes instead. Note that after the initial call to configure() changes to the canvas width and height will now take effect without the need to call configure() again.